Splitted generate_page into three elements: generate, generate_page and writer
